bui-themer: more files!
